requirements.txt
annotated-types==0.7.0
anyio==4.4.0
APScheduler==3.10.4
bcrypt==4.2.0
boto3==1.35.29
botocore==1.35.29
CacheControl==0.14.0
cachetools==5.4.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.0
et-xmlfile==1.1.0
fastapi==0.112.1
firebase-admin==6.5.0
google-api-core==2.19.1
google-api-python-client==2.147.0
google-auth==2.33.0
google-auth-httplib2==0.2.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.63.2
grpcio==1.65.4
grpcio-status==1.65.4
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==2.1.5
msgpack==1.1.0
mysql-connector-python==9.0.0
openpyxl==3.1.5
packaging==24.1
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.27.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.9.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-multipart==0.0.9
pytz==2024.1
requests==2.32.3
rsa==4.9
s3transfer==0.10.2
setuptools==72.2.0
six==1.16.0
sniffio==1.3.1
starlette==0.38.2
typing_extensions==4.12.2
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.30.6
websockets==13.0.1
